fixup! ARM: 7862/1: pcpu: replace __get_cpu_var_uses
The kvm/arm.c conflict resolution from the backport mistakenly brought
back part of dac288f7b38a7439502b77dabcdf8a9a5c4ae721 as well. Remove
that change.

Reviewed-on: http://git-master/r/409757
(cherry picked from commit 33f4e80db50ade81145bd631b1a2d48ba6fe89c6)
Reviewed-on: http://git-master/r/417588
GVS: Gerrit_Virtual_Submit
Tested-by: Bharat Nihalani <bnihalani@nvidia.com>

+++ arch/arm/kvm/arm.c
@@ -809,8 +809,8 @@ static void cpu_init_hyp_mode(void *dummy)
 	/* Switch from the HYP stub to our own HYP init vector */
 	__hyp_set_vectors(kvm_get_idmap_vector());
 
-	boot_pgd_ptr = kvm_mmu_get_boot_httbr();
-	pgd_ptr = kvm_mmu_get_httbr();
+	boot_pgd_ptr = (unsigned long long)kvm_mmu_get_boot_httbr();
+	pgd_ptr = (unsigned long long)kvm_mmu_get_httbr();
 	stack_page = __this_cpu_read(kvm_arm_hyp_stack_page);
 	hyp_stack_ptr = stack_page + PAGE_SIZE;
 	vector_ptr = (unsigned long)__kvm_hyp_vector;
